requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
PyMySQL==1.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
gunicorn==22.0.0
Werkzeug==3.0.3
WTForms==3.1.2
cryptography==42.0.8
